# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VenueNav API
  version: 1.0.0
  description: |
    REST API for multi-tenant indoor event maps: upload, processing jobs, graph/shops, routing, and search.
    All mutating admin routes require appropriate organization role (owner, editor).
servers:
  - url: https://api.venuenav.example.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgId:
      name: org_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EventId:
      name: event_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
security:
  - bearerAuth: []
paths:
...
  /organizations/{org_id}/events/{event_id}/venues:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Events]
      summary: List venues
      responses: { '200': { description: OK } }
    post:
      tags: [Events]
      summary: Create venue
      responses: { '201': { description: Created } }
...